# Task 1 - Implement a minimal version to download and parse the metadata files (redhat's Primary-xml file)