--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,4 @@
 .idea/
 metadata/
 db
+__pycache__/