--- a/requirements.txt
+++ b/requirements.txt
@@ -1,9 +1,17 @@
 certifi==2024.2.2
 charset-normalizer==3.3.2
+exceptiongroup==1.2.1
 greenlet==3.0.3
 idna==3.6
+iniconfig==2.0.0
+memory-profiler==0.61.0
+packaging==24.0
+pluggy==1.5.0
+psutil==5.9.8
 psycopg2-binary==2.9.9
+pytest==8.2.1
 requests==2.31.0
 SQLAlchemy==2.0.28
+tomli==2.0.1
 typing_extensions==4.10.0
 urllib3==2.2.1