## Description
@cbosdo @agraul
In this PR, we've made a minimal implementation of the process of downloading and parsing a package's metadata file (Primary-xml in our case).
## 1. File download
In order to make the download more memory efficient, we used **streams** to download and process the data in chunks. We also used **two different approaches to copy the content of each chunk into the target file**:
### 1.1 Default: writing directly to the file
```python
with requests.get(primary_xml_url, stream=True) as response:
            response.raise_for_status()  # check for status code
            with open(os.path.join(CACHE_DIR, primary_xml), 'wb') as f:
                for chunk in response.iter_content(chunk_size=8192):
                    # chunk_size to be efficiently determined
                    f.write(chunk)
                print(f"file {primary_xml} was downloaded successfully!")
```
### 1.2 Using shutil.copyfileobj
```python
    try:
        primary_xml_url = urljoin(url, primary_xml)
        response = requests.get(primary_xml_url, stream=True)
        with open(os.path.join(CACHE_DIR, primary_xml), 'wb') as out_file:
            shutil.copyfileobj(response.raw, out_file)
        print(f"file {primary_xml} was downloaded successfully!")
        response.close()
    except requests.exceptions.RequestException as e:
        print(f"Error downloading {primary_xml} file:", e)
```
For more information about shutil, please [shutil.copyfileobj](https://docs.python.org/3/library/shutil.html#shutil.copyfileobj)
## 2. File Parse (using SAX)
After downloading and extracting the Primary-xml gz file, we now have to parse it. To do it efficiently we used **SAX** instead of **DOM**, because it requires significantly less memory to process an XML document than the DOM.

Currently, the only thing we're doing is that we print, for each tag, its `name`, its `attributes`, its `Tag tree` (this is a variable that we defined and it can be useful to identify nested tags, and it will make it easier to give context to each tag), and its `value` if it has one.
Here's an example of the output:

```bash
current: [package] | attrs: {'type': 'rpm'}
Tag tree: package
------------------------------------------------------------
current: [name] | attrs: {}
Tag tree: package/name
------------------------------------------------------------
current: [name]
Value: gstreamer-plugins-bad
------------------------------------------------------------
...
------------------------------------------------------------
current: [rpm:entry] | attrs: {'name': 'gst-plugins-bad', 'flags': 'EQ', 'epoch': '0', 'ver': '1.22.0'}
Tag tree: package/format/rpm:provides/rpm:entry
------------------------------------------------------------
current: [rpm:entry] | attrs: {'name': 'gstreamer-plugins-bad', 'flags': 'EQ', 'epoch': '0', 'ver': '1.22.0', 'rel': 'lp155.3.4.1'}
Tag tree: package/format/rpm:provides/rpm:entry
```
## Benchmark
### Memory
We used a python library called `memory-profiler`, so we can have good idea on how much memory the download and the parsing are using. This is by decorating the method we want to benchmark with `@profile`.

In our case, we've decorated: 
```python 
@profile
def download_primary_xml_v2(url: str, primary_xml: str):
```
and
```python
@profile
def parse_primary_xml(file_path)
```
in order to visualize the used memory for both the download and the parse.
### Time
For the execution time, we used the linux `time` command on our program to have and idea about the total execution time. 

We might want to write a shell script that executes the program several times and return the average execution time. And clearing the memory cache after each iteration.
## Run
We've tested the program on the `https://download.opensuse.org/update/leap/15.5/oss/repodata/` repository, and more specifically on the `00c99a7e67dac325f1cbeeedddea3e4001a8f803c9bf43d9f50b5f1c756b0887-primary.xml.gz` file inside of it.

Here's the final run command:
```bash
time python3 -m memory_profiler implementation_1.py -u https://download.opensuse.org/update/leap/15.5/oss/repodata/
```
And here's the output:
```bash
Line #    Mem usage    Increment  Occurrences   Line Contents
=============================================================
    66     25.8 MiB     25.8 MiB           1   @profile
    67                                         def download_primary_xml_v2(url: str, primary_xml: str):
    68                                             """Download the primary-xml file from the given repository url"""
    69     25.8 MiB      0.0 MiB           1       try:
    70     25.8 MiB      0.0 MiB           1           primary_xml_url = urljoin(url, primary_xml)
    71     29.0 MiB      3.2 MiB           1           response = requests.get(primary_xml_url, stream=True)
    72     29.0 MiB      0.0 MiB           1           with open(os.path.join(CACHE_DIR, primary_xml), 'wb') as out_file:
    73     29.1 MiB      0.1 MiB           1               shutil.copyfileobj(response.raw, out_file)
    74     29.1 MiB      0.0 MiB           1           print(f"file {primary_xml} was downloaded successfully!")
    75     29.1 MiB      0.0 MiB           1           response.close()
    76                                             except requests.exceptions.RequestException as e:
    77                                                 print(f"Error downloading {primary_xml} file:", e)

...
Line #    Mem usage    Increment  Occurrences   Line Contents
=============================================================
    80     29.4 MiB     29.4 MiB           1   @profile
    81                                         def parse_primary_xml(file_path):
    82     29.4 MiB      0.0 MiB           1       handler = PrimaryXmlHandler()
    83     29.6 MiB      0.2 MiB           1       parser = xml.sax.make_parser()
    84                                             # parser.setFeature(xml.sax.handler.feature_namespaces, True)  # TODO Fix problem (probably for the root tag)
    85     29.6 MiB      0.0 MiB           1       parser.setContentHandler(handler)
    86                                         
    87     29.6 MiB      0.0 MiB           1       try:
    88     29.6 MiB      0.0 MiB           1           parser.parse(file_path)
    89     29.6 MiB      0.0 MiB           1       except Exception as e:
    90     29.6 MiB      0.0 MiB           1           print("Error parsing file: ", e)
...
Line #    Mem usage    Increment  Occurrences   Line Contents
=============================================================
   136     25.8 MiB     25.8 MiB           1   @profile
   137                                         def download_and_parse_metadata(program_args):
   138     25.8 MiB      0.0 MiB           1       if not program_args.url:
   139                                                 print('Error: target url not defined!')
   140                                                 raise ValueError('Target url was not provided. Please run python prog -u https://your-repository-url.com')
   141                                         
   142     25.8 MiB      0.0 MiB           1       repo_url = program_args.url
   143                                             # print(repo_url, end='')
   144     29.1 MiB      3.4 MiB           1       download_primary_xml_v2(repo_url, PRIMARY_XML)
   145     29.4 MiB      0.2 MiB           1       extract_file(CACHE_DIR, PRIMARY_XML)
   146     29.6 MiB      0.2 MiB           1       parse_primary_xml(os.path.join(CACHE_DIR, PRIMARY_XML.rstrip(".gz")))
...
real    0m4.559s
user    0m2.240s
sys     0m0.054s
```

## TODO
- **chunk_size**: I think we should determine the best `chunk_size` for each processed file. Maybe we can make a formula that consider the **file size** and the current **free memory** on the system.
- **Signature**: we're still not checking for the signature after the download, so we should do that.
- **// Processing**: is it worth, and possible, to use parallel processing at this stage ?
- **namespace mode**: Currently the namespace mode is not active; this means that the `def startElementNS(..)` and `def endElementNS(..)` are not working. We can activate them using `parser.setFeature(xml.sax.handler.feature_namespaces, True)`, but we still have a minor parsing issue (probably related to the root element).